Move Predicate to Frontend
