+++ hs-bindgen/src-internal/HsBindgen/Frontend/Pass/Parse.hs
@@ -4,12 +4,12 @@ module HsBindgen.Frontend.Pass.Parse (parseDecls) where
 import Clang.HighLevel qualified as HighLevel
 import Clang.LowLevel.Core
 
-import HsBindgen.C.Predicate qualified as Predicate
 import HsBindgen.Frontend.Analysis.IncludeGraph (IncludeGraph)
 import HsBindgen.Frontend.AST.Internal qualified as C
 import HsBindgen.Frontend.Pass.Parse.Decl
 import HsBindgen.Frontend.Pass.Parse.Decl.Monad qualified as ParseDecl
 import HsBindgen.Frontend.Pass.Parse.IsPass
+import HsBindgen.Frontend.Predicate qualified as Predicate
 import HsBindgen.Frontend.ProcessIncludes (GetMainHeader)
 import HsBindgen.Frontend.RootHeader (RootHeader)
 import HsBindgen.Util.Tracer
