+++ hs-bindgen/src/HsBindgen/Common.hs
@@ -96,9 +96,9 @@ import Clang.Args qualified as Args
 import Clang.Paths qualified as Paths
 
 import HsBindgen.BindingSpec qualified as BindingSpec
-import HsBindgen.C.Predicate qualified as Predicate
 import HsBindgen.Config qualified as Config
 import HsBindgen.Frontend.Pass.Select.IsPass qualified as Select
+import HsBindgen.Frontend.Predicate qualified as Predicate
 import HsBindgen.Frontend.RootHeader qualified as RootHeader
 import HsBindgen.Hs.AST qualified as Hs
 import HsBindgen.Hs.Translation qualified as Hs
