+++ hs-bindgen/test/hs-bindgen/Test/HsBindgen/Golden.hs
@@ -11,12 +11,12 @@ import Clang.Enum.Simple
 import Clang.LowLevel.Core
 import Clang.Version
 import HsBindgen.BindingSpec qualified as BindingSpec
-import HsBindgen.C.Predicate
-  (DeclPredicate(..), HeaderPathPredicate(..), Predicate(..))
 import HsBindgen.Config
 import HsBindgen.Frontend.AST.Internal qualified as C
 import HsBindgen.Frontend.Naming qualified as C
 import HsBindgen.Frontend.Pass.Select.IsPass
+import HsBindgen.Frontend.Predicate (DeclPredicate (..),
+                                     HeaderPathPredicate (..), Predicate (..))
 import HsBindgen.TraceMsg
 
 import Test.Common.HsBindgen.TracePredicate
